PIZ-1: code refactor and changing file structure
